model hub improvements. different structure of calling methods, new models and folder structure
