+++ analysis/bach_open_taxonomy/bach_open_taxonomy/__init__.py
@@ -3,5 +3,7 @@ Copyright 2021 Objectiv B.V.
 """
 __version__ = '0.0.1'
 
+from bach_open_taxonomy.objectivframe import ObjectivFrame
 from bach_open_taxonomy.series import *
+from bach_open_taxonomy.modelhub import *
 from bach_open_taxonomy.stack.util import basic_feature_model
