+++ analysis/bach_open_taxonomy/bach_open_taxonomy/modelhub/__init__.py
@@ -1,0 +1,3 @@
+"""
+Copyright 2021 Objectiv B.V.
+"""
