+++ analysis/bach_open_taxonomy/bach_open_taxonomy/modelhub/aggregate.py
@@ -1,0 +1,101 @@
+"""
+Copyright 2021 Objectiv B.V.
+"""
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from bach.series import SeriesBoolean
+
+
+class Aggregate:
+    """
+    Models that return aggregated data in some form from the original ObjectivFrame.
+
+    Methods in this class can be filtered with the filter parameter, which always takes SeriesBoolean. The
+    ModelHub can also create specific commonly used filters with methods from
+    :py:attr:`bach_open_taxonomy.ModelHub.map`.
+    """
+
+    def __init__(self, df):
+        self._df = df
+
+    def _generic_aggregation(self, time_aggregation, column, filter, name):
+        if not time_aggregation:
+            time_aggregation = self._df._time_aggregation
+        gb = self._df.moment.dt.sql_format(time_aggregation) if time_aggregation else None
+        df = self._df.copy_override()
+        if filter:
+            df['_filter'] = filter
+            if filter.expression.has_windowed_aggregate_function:
+                df = df.materialize()
+            df = df[df._filter]
+
+            name += '_' + filter.name
+
+        series = df.groupby(gb)[column].nunique()
+        return series.copy_override(name=name)
+
+    def unique_users(self, time_aggregation: str = None, filter: 'SeriesBoolean' = None):
+        """
+        Calculate the unique users in the ObjectivFrame.
+
+        Use any template for aggreation from: https://www.postgresql.org/docs/14/functions-formatting.html
+        ie. ``time_aggregation=='YYYY-MM-DD'`` aggregates by date.
+
+        :param time_aggregation: if None, it uses the time_aggregation set in ObjectivFrame.
+        :param filter: the output of this model is only based on the rows for which the filter is True.
+        """
+
+        return self._generic_aggregation(time_aggregation=time_aggregation,
+                                         column='user_id',
+                                         filter=filter,
+                                         name='unique_users')
+
+    def unique_sessions(self, time_aggregation: str = None, filter: 'SeriesBoolean' = None):
+        """
+        Calculate the unique sessions in the ObjectivFrame.
+
+        Use any template for aggreation from: https://www.postgresql.org/docs/14/functions-formatting.html
+        ie. ``time_aggregation=='YYYY-MM-DD'`` aggregates by date.
+
+        :param time_aggregation: if None, it uses the time_aggregation set in ObjectivFrame.
+        :param filter: the output of this model is only based on the rows for which the filter is True.
+        """
+
+        return self._generic_aggregation(time_aggregation=time_aggregation,
+                                         column='session_id',
+                                         filter=filter,
+                                         name='unique_sessions')
+
+    def session_duration(self, time_aggregation: str = None):
+        # calculate duration of each session
+        df = self._df.copy_override()
+        if not time_aggregation:
+            time_aggregation = self._df._time_aggregation
+        gb = ['session_id']
+        if time_aggregation:
+            df[time_aggregation] = df.moment.dt.sql_format(time_aggregation)
+            gb = ['session_id', time_aggregation]
+
+        session_duration = df.groupby(gb).aggregate(
+            {'moment': ['min', 'max']})
+
+        session_duration['session_duration'] = session_duration['moment_max'] - session_duration['moment_min']
+
+        # check which sessions have duration of zero and filter these out, as they are bounces
+        session_duration = session_duration[(session_duration['session_duration'] > '0')]
+
+        return session_duration.groupby(time_aggregation).aggregate(
+            {'session_duration': 'mean'})
+
+    def frequency(self):
+        # number of total sessions per user
+        total_sessions_user = self._df.groupby(['user_id']).aggregate({'session_id': 'nunique'})
+
+        # calculate frequency
+        frequency = total_sessions_user.groupby(['session_id_nunique']).aggregate({'user_id': 'nunique'})
+
+        # add total users and calculate share per number of sessions
+        frequency['share_of_users'] = frequency['user_id_nunique'] / self._df['user_id'].nunique()
+
+        return frequency
